services/accounts-service/src/main/java/com/secure_sentinel/accounts/accounts_service/model/enums/AccountType.java
package com.secure_sentinel.accounts.accounts_service.model.enums;

import com.fasterxml.jackson.annotation.JsonCreator;
import com.fasterxml.jackson.annotation.JsonValue;

public enum AccountType {
    CHECKING("Checking"),
    SAVINGS("Savings"),
    CREDIT("Credit"),
    LOAN("Loan");

    private final String value;

    AccountType(String value) {
        this.value = value;
    }

    @JsonValue
    public String getValue() {
        return value;
    }

    @JsonCreator
    public static AccountType forValue(String value) {
        for (AccountType type : values()) {
            if (type.getValue().equalsIgnoreCase(value)) {
                return type;
            }
        }
        throw new IllegalArgumentException("Invalid Account type: " + value);
    }
}